fix: soporte completo de decimales en cantidades (importar, POS, compras, devoluciones)
- Importar Excel: nueva función parseQty() que interpreta "212,5" como 212.5
- POS carrito: cantidad editable directamente (click para escribir decimales)
- Compras backend: parseFloat en lugar de parseInt para cantidades y stock
- Devoluciones: parseFloat para cantidades devueltas

+++ src/features/pos/hooks/useReturn.js
@@ -46,7 +46,7 @@ export function useReturn({ cuentas, fetchData }) {
             .filter(([_, qty]) => qty > 0)
             .map(([itemVentaId, cantidad]) => ({
                 itemVentaId: parseInt(itemVentaId),
-                cantidad: parseInt(cantidad)
+                cantidad: parseFloat(cantidad)
             }));
 
         if (itemsToReturn.length === 0) {
